IncomeCalculation
Made classes for buildings, building data, enums, and a game manager. Put methods in the game manager that can calculate the player's income just based on the buildings it knows about, and automatically adjusts the income when a building is built or removed, or a connection is added between two buildings.