LevelBuilder
Made a basic level builder which will create a new scriptable object containing the starting cash, goal cash, and mission time automatically when the level resource is saved. Currently how I'm placing buildings on planets is to place them nearby and start the game to run the function that places a building on a planet surface, since colliders don't work in editor.