FixedCollision/GameSpeed/Science
Fixed planet collision so it only gets called from the bigger planet, so both planets only get destroyed if they're the same size. Added the ability to speed up and slow down the game with the P and O keys respectivly. Added a new resource called science, that better buildings require so the player has to decide which to focus on between science and money at various points in the game. (Could be called advanced materials instead.