EditableLevelsFolder
Made it so we save the levels as prefabs to go back and edit them easier as other prefabs in the game change. Also made it so planets can never return total mass less than 1 if we ever wanted to do 0, or negative mass buildings that reduce the risk of planet crashing.