MoveToolUpdates
Made it so the move tool now no longer blocks the hover and selection building highlights, which makes it easier for the player to tell which building they're moving. Also, made it so the player can change which building they're moving by clicking on a new building, even if one is already selected, and after successfully moving a building, the currently being moved building is deselected, without exiting the move tool.